# [Refactor] `void-dom-elements-no-children`: improve performance