--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -14,7 +14,9 @@ This change log adheres to standards from [Keep a CHANGELOG](http://keepachangel
 
 ### Changed
 * [Docs] [`jsx-newline`]: Fix minor spelling error on rule name ([#2974][] @DennisSkoko)
+* [Refactor] [`void-dom-elements-no-children`]: improve performance
 
+[#2977]: https://github.com/yannickcr/eslint-plugin-react/pull/2977
 [#2975]: https://github.com/yannickcr/eslint-plugin-react/pull/2975
 [#2974]: https://github.com/yannickcr/eslint-plugin-react/pull/2974
 [#2972]: https://github.com/yannickcr/eslint-plugin-react/pull/2972