--- a/lib/util/Components.js
+++ b/lib/util/Components.js
@@ -393,25 +393,29 @@ function componentRule(rule, context) {
      * @returns {Boolean} True if createElement called from pragma
      */
     isCreateElement(node) {
-      const calledOnPragma = (
+      // match `React.createElement()`
+      if (
         node
         && node.callee
         && node.callee.object
         && node.callee.object.name === pragma
         && node.callee.property
         && node.callee.property.name === 'createElement'
-      );
+      ) {
+        return true;
+      }
 
-      const calledDirectly = (
+      // match `createElement()`
+      if (
         node
         && node.callee
         && node.callee.name === 'createElement'
-      );
-
-      if (this.isDestructuredFromPragmaImport('createElement')) {
-        return calledDirectly || calledOnPragma;
+        && this.isDestructuredFromPragmaImport('createElement')
+      ) {
+        return true;
       }
-      return calledOnPragma;
+
+      return false;
     },
 
     /**