closes #1126.

The rule spends lots of time in`utils.isCreateElement` -> `isDestructuredFromPragmaImport`.  This pr reduces that expensive function call by short-circuiting it when possible.

Benchmarks:
```
Before this pr:
$ TIMING=50 npx eslint . | grep react/void-dom
react/void-dom-elements-no-children             |   263.673 |     3.6%
react/void-dom-elements-no-children             |   256.981 |     3.4%
react/void-dom-elements-no-children             |   230.990 |     3.3%
react/void-dom-elements-no-children             |   235.888 |     3.2%
react/void-dom-elements-no-children             |   251.152 |     3.5%

After this pr:
react/void-dom-elements-no-children             |   106.662 |     1.6%
react/void-dom-elements-no-children             |   106.205 |     1.6%
react/void-dom-elements-no-children             |   106.061 |     1.6%
react/void-dom-elements-no-children             |   103.666 |     1.6%
react/void-dom-elements-no-children             |   103.363 |     1.6%
```